fix the problem where moving a card and then deleting it causes an error
This happens in a non-editable editor if you move the bottom card to the
top and then remove it. Because `_readRangeFromDOM` doesn't do anything
if the editor is non-editable, this means it may get stuck with a range
that is pointing at a card that has been removed from the post.
This causes the error seen when doing the `post.walkAllLeafSections`
later in the snapshot code. The post is undefined because it is read off
the card, but the card that the range points at has been removed.