Grouped undo and redo statements. refs #502
We created three types of events can be stored in the edit history. Markup
insertions, deletions, and all other actions. When a markup insertion or
deletions happens within Xms of a previous edit of the same type, those
edits may be "grouped" for undo/redo. Practically, it means the pending
snapshot on history is stomped with the more recent snapshot.